Made changes discussed in pull request.
