+++ LibationLauncher/LibationLauncher.csproj
@@ -13,7 +13,7 @@
     <!-- <PublishSingleFile>true</PublishSingleFile> -->
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
 
-    <Version>4.4.0.64</Version>
+    <Version>4.4.0.120</Version>
   </PropertyGroup>
 
   <ItemGroup>
